GUI beginning
This is the first iteration of a compatible GUI class - there is still some translation that I will do from a more CL interface to a GU interface format (won't use println and such), but I wanted you guys to be able to see some of the work! Don't try to run this with the main yet, I'll put it in myself when it is better runnable. :) I'll also delete this later once  the final iteration is done.